requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.3
langchain==0.3.23
langchain-community==0.3.21
openai==1.75.0
jinja2==3.1.6
markdown2==2.4.13
python-multipart==0.0.20
langchain-openai==0.3.14

